Check for existing file before deleting

+++ link.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 
-rm ~/.vimrc
+if [[ -e "$HOME/.vimrc" ]]; then
+  rm ~/.vimrc
+fi
 ln -rs vim/.vimrc ~/.vimrc
-rm ~/.tmux.conf
+if [[ -e "$HOME/.tmux.conf" ]]; then
+  rm ~/.tmux.conf
+fi
 ln -rs tmux/.tmux.conf ~/.tmux.conf
 exit 0
